Refactor worker implementation
