+++ cmd/capabilities/demo.sh
