+++ deployments/capabilities/dispatcher/Dockerfile.dispatcher
@@ -8,7 +8,7 @@ FROM golang:1.15 as builder
 WORKDIR /app
 # Copy local code to the container image.
 COPY . ./
-RUN CGO_ENABLED=0 GOOS=linux go build -v -o dispatcher ./cmd/capabilities/dispatcher/dispatcher.go
+RUN CGO_ENABLED=0 GOOS=linux go build -v -o dispatcher ./cmd/capabilities/dispatcher-server/main.go
 
 # Use the official Alpine image for a lean production container.
 # https://hub.docker.com/_/alpine
