+++ queue.yaml
@@ -1,1 +1,1 @@
-cmd/capabilities/task-queue/registry-task-queue.yaml
+deployments/capabilities/task-queue/registry-task-queue.yaml
